Use scripts for loading sky
After the Vegetation I now made Lua scripts for inserting the parameters
of the SkyManager's objects. There were only a few and since there's
only one sky I kept things rather simple. However, in the future the
SkyManager should also do stuff like clouds and so, so there is no
reason not to use Lua scripts here as well.

+++ src/core/veg.hpp
@@ -57,7 +57,7 @@ namespace ch
 				}
 
 				/** Add a LOD for a type of vegetation. Use 'double' for tileSize/maxRange because Lua likes that. */
-				void addLOD(std::string name, double tileSize, std::string meshLoc, std::string difTexLoc, bool alphaChannel, std::string normTexLoc, int maxObjects, double maxRange);
+				void addLOD(std::string name, float tileSize, std::string meshLoc, std::string difTexLoc, bool alphaChannel, std::string normTexLoc, int maxObjects, float maxRange);
 				/** Add all LODs to their TiledHordes. */
 				void addToTiledHorde(void) { for(std::map<std::string, veg::VegetationTiledHorde*>::iterator it = hordemap.begin(); it != hordemap.end(); it++) it->second->addToTiledHorde(); }
 				/** Register Lua functions. */
